Final refactor (paid off technical debt)
