+++ src/renderer/components/App.css
@@ -43,6 +43,4 @@ body {
 }
 
 .App__body {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
 }
