--- src/renderer/components/HelmetProject.css
@@ -1,63 +1,0 @@
-.Project {
-    width: 24em;
-    padding-left: 0.75em;
-}
-
-.Project__heading {
-    font-size: 1.5em;
-    margin: 0 0 1em 0;
-    font-weight: normal;
-}
-
-.Project__saved-scenarios {
-    display: table;
-}
-
-.Project__scenario {
-    display: table-row;
-}
-
-.Project__scenario-run-btn,
-.Project__scenario-name-config-btn {
-    display: table-cell;
-    margin: 0.25em;
-    line-height: 1.75em;
-}
-
-.Project__scenario-run-btn--active {
-    background: mediumseagreen;
-}
-
-.Project__scenario-name-config-btn--active {
-    background: lightslategrey;
-}
-
-.Project__separator {
-    margin: 1em 0 0 0;
-}
-
-.Project__start-stop-controls {
-}
-
-.Project__start-stop-description {
-}
-
-.Project__run-status-percentage-ready {
-    margin: 1em 0;
-    border: 1px solid silver;
-    padding: 4px;
-}
-
-.Project__run-status-scenario-ready {
-}
-
-.Project__run-status-logfile-link {
-}
-
-.Project__start-stop-scenarios {
-    color: mediumseagreen;
-    font-weight: bold;
-}
-
-.Project__start-stop-btn {
-}
