+++ src/renderer/components/HelmetProject/HelmetProject.css
@@ -1,0 +1,4 @@
+.Project {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+}
