+++ src/renderer/index.html
@@ -13,14 +13,18 @@
     <!-- React components -->
     <script type="text/babel" src="components/App.jsx"></script>
     <link rel="stylesheet" type="text/css" href="components/App.css">
-    <script type="text/babel" src="components/Settings.jsx"></script>
-    <link rel="stylesheet" type="text/css" href="components/Settings.css">
-    <script type="text/babel" src="components/HelmetProject.jsx"></script>
-    <link rel="stylesheet" type="text/css" href="components/HelmetProject.css">
-    <script type="text/babel" src="components/HelmetScenario.jsx"></script>
-    <link rel="stylesheet" type="text/css" href="components/HelmetScenario.css">
-    <script type="text/babel" src="components/RunLog.jsx"></script>
-    <link rel="stylesheet" type="text/css" href="components/RunLog.css">
+    <script type="text/babel" src="components/Settings/Settings.jsx"></script>
+    <link rel="stylesheet" type="text/css" href="components/Settings/Settings.css">
+    <script type="text/babel" src="components/HelmetProject/HelmetProject.jsx"></script>
+    <link rel="stylesheet" type="text/css" href="components/HelmetProject/HelmetProject.css">
+    <script type="text/babel" src="components/HelmetProject/HelmetScenario/HelmetScenario.jsx"></script>
+    <link rel="stylesheet" type="text/css" href="components/HelmetProject/HelmetScenario/HelmetScenario.css">
+    <script type="text/babel" src="components/HelmetProject/Runtime/Runtime.jsx"></script>
+    <link rel="stylesheet" type="text/css" href="components/HelmetProject/Runtime/Runtime.css">
+    <script type="text/babel" src="components/HelmetProject/Runtime/RunStatus/RunStatus.jsx"></script>
+    <link rel="stylesheet" type="text/css" href="components/HelmetProject/Runtime/RunStatus/RunStatus.css">
+    <script type="text/babel" src="components/HelmetProject/RunLog/RunLog.jsx"></script>
+    <link rel="stylesheet" type="text/css" href="components/HelmetProject/RunLog/RunLog.css">
     <script type="text/javascript">
         const {webFrame} = require('electron');
         // Disable content zooming (it seems prone to glitching on Windows). https://github.com/electron/electron/issues/15496
